Remove Angr from requirements.txt file.
This should be installed manually using Python3, as Angr dropped support for Python2 whilst on the other hand templite only supports Python2 which uses to compile this project.
That being said, for bootstrapping the project use Python2 and for Angr use Python3.
This is not as convenient, but when using virtualenv this is a simple switch.